mkimg.sh: Declare /usr/portage and /var/lib/layman mountpoints.
This allows us to ship a "development" image that has all the binaries
and Portage tree embedded, and a "runtime" image without.

To install a package, we spin up the "development" image into a running
container and let it build any additional packages needed, then we can
spin up the runtime using `--volumes-from` pointing at the development
container before running `emerge -K` to install them.

After that is done, `docker commit` will commit that instance of the
runtime image to a new image which is our bundled package container.